refactor: move policy related models to separate directory
